Extract Server implementation from pool module
In order to track stats on the Server, the pool had to stop making deep
clones of the Server. The actual Server implementation is now hidden
away behind Arc<RwLock<T>>. This is similar to how the Pool works.

Fixes #66
